[ETCM-518] Refactor EthService by adding EthBlocksSevice - service that handles RPC calls related to blocks
