+++ src/main/scala/io/iohk/ethereum/nodebuilder/NodeBuilder.scala
@@ -401,7 +401,7 @@ trait EthServiceBuilder {
   )
 }
 
-trait MiningServiceBuilder {
+trait EthMiningServiceBuilder {
   self: BlockchainBuilder
     with LedgerBuilder
     with JSONRpcConfigBuilder
@@ -410,7 +410,7 @@ trait MiningServiceBuilder {
     with PendingTransactionsManagerBuilder
     with TxPoolConfigBuilder =>
 
-  lazy val miningService = new EthMiningService(
+  lazy val ethMiningService = new EthMiningService(
     blockchain,
     ledger,
     jsonRpcConfig,
@@ -419,6 +419,12 @@ trait MiningServiceBuilder {
     pendingTransactionsManager,
     txPoolConfig.getTransactionFromPoolTimeout
   )
+}
+
+trait EthBlocksServiceBuilder {
+  self: BlockchainBuilder with BlockchainConfigBuilder with LedgerBuilder =>
+
+  lazy val ethBlocksService = new EthBlocksService(blockchain, ledger, blockchainConfig)
 }
 
 trait PersonalServiceBuilder {
@@ -503,6 +509,8 @@ trait JSONRpcControllerBuilder {
     with EthServiceBuilder
     with EthProofServiceBuilder
     with MiningServiceBuilder
+    with EthMiningServiceBuilder
+    with EthBlocksServiceBuilder
     with NetServiceBuilder
     with PersonalServiceBuilder
     with DebugServiceBuilder
@@ -520,6 +528,8 @@ trait JSONRpcControllerBuilder {
       web3Service,
       netService,
       ethService,
+      ethMiningService,
+      ethBlocksService,
       personalService,
       testService,
       debugService,
@@ -533,8 +543,8 @@ trait JSONRpcControllerBuilder {
 }
 
 trait JSONRpcHealthcheckerBuilder {
-  this: NetServiceBuilder with EthServiceBuilder =>
-  lazy val jsonRpcHealthChecker: JsonRpcHealthChecker = new NodeJsonRpcHealthChecker(netService, ethService)
+  this: NetServiceBuilder with EthBlocksServiceBuilder =>
+  lazy val jsonRpcHealthChecker: JsonRpcHealthChecker = new NodeJsonRpcHealthChecker(netService, ethBlocksService)
 }
 
 trait JSONRpcHttpServerBuilder {
@@ -699,6 +709,8 @@ trait Node
     with EthServiceBuilder
     with EthProofServiceBuilder
     with MiningServiceBuilder
+    with EthMiningServiceBuilder
+    with EthBlocksServiceBuilder
     with NetServiceBuilder
     with PersonalServiceBuilder
     with DebugServiceBuilder
