+++ src/test/scala/io/iohk/ethereum/jsonrpc/CheckpointingJRCSpec.scala
@@ -208,16 +208,19 @@ class CheckpointingJRCSpec
     val personalService = mock[PersonalService]
     val debugService = mock[DebugService]
     val ethService = mock[EthService]
+    val ethMiningService = mock[EthMiningService]
+    val ethBlocksService = mock[EthBlocksService]
     val qaService = mock[QAService]
     val checkpointingService = mock[CheckpointingService]
     val mantisService = mock[MantisService]
-    val miningService = mock[EthMiningService]
 
     val jsonRpcController =
       new JsonRpcController(
         web3Service,
         netService,
         ethService,
+        ethMiningService,
+        ethBlocksService,
         personalService,
         None,
         debugService,
