+++ src/test/scala/io/iohk/ethereum/jsonrpc/JsonRpcControllerEthSpec.scala
@@ -5,11 +5,10 @@ import akka.testkit.TestKit
 import akka.util.ByteString
 import io.iohk.ethereum.blockchain.sync.SyncProtocol
 import io.iohk.ethereum.blockchain.sync.SyncProtocol.Status.Progress
-import io.iohk.ethereum.consensus.Consensus
 import io.iohk.ethereum.consensus.blocks.{PendingBlock, PendingBlockAndState}
-import io.iohk.ethereum.consensus.validators.SignedTransactionValidator
 import io.iohk.ethereum.crypto.kec256
 import io.iohk.ethereum.domain._
+import io.iohk.ethereum.jsonrpc.EthBlocksService.{GetUncleCountByBlockHashResponse, GetUncleCountByBlockNumberResponse}
 import io.iohk.ethereum.jsonrpc.EthService._
 import io.iohk.ethereum.jsonrpc.FilterManager.LogFilterLogs
 import io.iohk.ethereum.jsonrpc.PersonalService._
@@ -155,6 +154,12 @@ class JsonRpcControllerEthSpec
   }
 
   it should "handle eth_getBlockByNumber request" in new JsonRpcControllerFixture {
+    (() => validators.signedTransactionValidator)
+      .expects()
+      .returns(null)
+      .anyNumberOfTimes()
+    (() => ledger.consensus).expects().returns(consensus)
+
     val blockToRequest = Block(Fixtures.Blocks.Block3125369.header, Fixtures.Blocks.Block3125369.body)
     val blockWeight = ChainWeight.zero.increase(blockToRequest.header)
 
@@ -176,6 +181,12 @@ class JsonRpcControllerEthSpec
   }
 
   it should "handle eth_getBlockByNumber request (block with treasuryOptOut)" in new JsonRpcControllerFixture {
+    (() => validators.signedTransactionValidator)
+      .expects()
+      .returns(null)
+      .anyNumberOfTimes()
+    (() => ledger.consensus).expects().returns(consensus)
+
     val blockToRequest = blockWithTreasuryOptOut
     val blockWeight = ChainWeight.zero.increase(blockToRequest.header)
 
@@ -197,6 +208,12 @@ class JsonRpcControllerEthSpec
   }
 
   it should "handle eth_getBlockByNumber request (block with checkpoint)" in new JsonRpcControllerFixture {
+    (() => validators.signedTransactionValidator)
+      .expects()
+      .returns(null)
+      .anyNumberOfTimes()
+    (() => ledger.consensus).expects().returns(consensus)
+
     val blockToRequest = blockWithCheckpoint
     val blockWeight = ChainWeight.zero.increase(blockToRequest.header)
 
@@ -243,6 +260,12 @@ class JsonRpcControllerEthSpec
   }
 
   it should "handle eth_getUncleByBlockNumberAndIndex request" in new JsonRpcControllerFixture {
+    (() => validators.signedTransactionValidator)
+      .expects()
+      .returns(null)
+      .anyNumberOfTimes()
+    (() => ledger.consensus).expects().returns(consensus)
+
     val uncle = Fixtures.Blocks.DaoForkBlock.header
     val blockToRequest = Block(Fixtures.Blocks.Block3125369.header, BlockBody(Nil, Seq(uncle)))
 
@@ -428,7 +451,9 @@ class JsonRpcControllerEthSpec
 
   it should "eth_call" in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
     (mockEthService.call _).expects(*).returning(Task.now(Right(CallResponse(ByteString("asd")))))
 
@@ -451,7 +476,8 @@ class JsonRpcControllerEthSpec
 
   it should "eth_estimateGas" in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
     (mockEthService.estimateGas _)
       .expects(*)
@@ -486,7 +512,8 @@ class JsonRpcControllerEthSpec
 
   it should "eth_getCode" in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
     (mockEthService.getCode _)
       .expects(*)
@@ -506,9 +533,10 @@ class JsonRpcControllerEthSpec
 
   it should "eth_getUncleCountByBlockNumber" in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
-    (mockEthService.getUncleCountByBlockNumber _)
+    (mockEthBlocksService.getUncleCountByBlockNumber _)
       .expects(*)
       .returning(Task.now(Right(GetUncleCountByBlockNumberResponse(2))))
 
@@ -525,9 +553,10 @@ class JsonRpcControllerEthSpec
 
   it should "eth_getUncleCountByBlockHash " in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
-    (mockEthService.getUncleCountByBlockHash _)
+    (mockEthBlocksService.getUncleCountByBlockHash _)
       .expects(*)
       .returning(Task.now(Right(GetUncleCountByBlockHashResponse(3))))
 
@@ -559,7 +588,8 @@ class JsonRpcControllerEthSpec
 
   it should "eth_getBalance" in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
     (mockEthService.getBalance _)
       .expects(*)
@@ -579,7 +609,8 @@ class JsonRpcControllerEthSpec
 
   it should "return error with custom error in data in eth_balance" in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
     (mockEthService.getBalance _)
       .expects(*)
@@ -599,7 +630,8 @@ class JsonRpcControllerEthSpec
 
   it should "eth_getStorageAt" in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
     (mockEthService.getStorageAt _)
       .expects(*)
@@ -646,7 +678,8 @@ class JsonRpcControllerEthSpec
 
   it should "eth_newFilter" in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
     (mockEthService.newFilter _)
       .expects(*)
@@ -670,7 +703,8 @@ class JsonRpcControllerEthSpec
 
   it should "eth_newBlockFilter" in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
     (mockEthService.newBlockFilter _)
       .expects(*)
@@ -689,7 +723,8 @@ class JsonRpcControllerEthSpec
 
   it should "eth_newPendingTransactionFilter" in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
     (mockEthService.newPendingTransactionFilter _)
       .expects(*)
@@ -706,7 +741,8 @@ class JsonRpcControllerEthSpec
 
   it should "eth_uninstallFilter" in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
     (mockEthService.uninstallFilter _)
       .expects(*)
@@ -723,7 +759,8 @@ class JsonRpcControllerEthSpec
 
   it should "eth_getFilterChanges" in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
     (mockEthService.getFilterChanges _)
       .expects(*)
@@ -865,7 +902,8 @@ class JsonRpcControllerEthSpec
 
   it should "eth_getFilterLogs" in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
     (mockEthService.getFilterLogs _)
       .expects(*)
@@ -894,7 +932,9 @@ class JsonRpcControllerEthSpec
 
   it should "eth_getLogs" in new JsonRpcControllerFixture {
     val mockEthService = mock[EthService]
-    override val jsonRpcController = newJsonRpcController(mockEthService)
+    val mockEthBlocksService = mock[EthBlocksService]
+
+    override val jsonRpcController = newJsonRpcController(mockEthService, mockEthBlocksService)
 
     (mockEthService.getLogs _)
       .expects(*)
