+++ src/test/scala/io/iohk/ethereum/jsonrpc/QaJRCSpec.scala
@@ -244,9 +244,10 @@ class QaJRCSpec
     val personalService = mock[PersonalService]
     val debugService = mock[DebugService]
     val ethService = mock[EthService]
+    val ethMiningService = mock[EthMiningService]
+    val ethBlocksService = mock[EthBlocksService]
     val checkpointingService = mock[CheckpointingService]
     val mantisService = mock[MantisService]
-    val miningService = mock[EthMiningService]
     val qaService = mock[QAService]
 
     val jsonRpcController =
@@ -254,6 +255,8 @@ class QaJRCSpec
         web3Service,
         netService,
         ethService,
+        ethMiningService,
+        ethBlocksService,
         personalService,
         None,
         debugService,
@@ -265,7 +268,7 @@ class QaJRCSpec
         config
       )
 
-    val mineBlocksReq = MineBlocksRequest(1, true, None)
+    val mineBlocksReq = MineBlocksRequest(1, withTransactions = true, None)
 
     val mineBlocksRpcRequest = JsonRpcRequest(
       "2.0",
